requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
openai>=1.3.0
supabase>=2.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
greenlet>=3.0.0
asyncpg>=0.29.0
alembic>=1.13.0
yt-dlp>=2023.12.0
ffmpeg-python>=0.2.0
boto3>=1.34.0
rq>=1.15.0
redis>=5.0.0
upstash-redis>=1.0.0
python-multipart>=0.0.9
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
scipy>=1.11.0
pillow>=10.0.0
opencv-python>=4.8.0
torchaudio>=2.0.0
soundfile>=0.12.0
pydub>=0.25.0
audioop-lts>=0.2.2
scikit-learn>=1.3.0
spacy>=3.7.0
torch sentence-transformers
torch torchaudio
torch transformers
torch sentence-transformers
torch transformers
